pyproject.toml
[project]
name = "banking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
	"grpcio>=1.80.0",
	"protobuf>=4.25.0",
	"rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
	"grpcio-tools>=1.80.0",
	"pytest>=8.0.0",
]
